add access_token authentication

+++ src/main/java/doitincloud/rdbcache/services/ExpireOps.java
@@ -9,7 +9,6 @@ package doitincloud.rdbcache.services;
 import doitincloud.rdbcache.configs.PropCfg;
 import doitincloud.commons.Utils;
 import doitincloud.rdbcache.configs.AppCtx;
-import doitincloud.rdbcache.controllers.RdbcacheApis;
 import doitincloud.rdbcache.models.KeyInfo;
 import doitincloud.rdbcache.models.KvPair;
 import doitincloud.rdbcache.models.StopWatch;
@@ -34,6 +33,8 @@ import org.slf4j.LoggerFactory;
 
 import javax.annotation.PostConstruct;
 import java.util.Collections;
+import java.util.List;
+import java.util.Map;
 
 @Service
 public class ExpireOps {
@@ -135,6 +136,7 @@ public class ExpireOps {
     public void setExpireKey(Context context, KvPair pair, KeyInfo keyInfo) {
 
         try {
+
             String key = pair.getId();
             String type = pair.getType();
 
@@ -142,6 +144,11 @@ public class ExpireOps {
 
             String expire = keyInfo.getExpire();
             String expKey = eventPrefix + "::" + type + ":" + key;
+
+            boolean noOps = keyInfo.isNoOps();
+            if (noOps) {
+                expKey += "::" + keyInfo.getQueryKey() + "/" + expire;
+            }
 
             StopWatch stopWatch = context.startStopWatch("redis", "scriptExecutor.execute");
             Long result = scriptExecutor.execute(set_expire_key_script,
@@ -151,7 +158,7 @@ public class ExpireOps {
             if (result != 1) {
                 keyInfo.restoreExpire();
             }
-            if (keyInfo.getIsNew()) {
+            if (!noOps && keyInfo.getIsNew()) {
                 AppCtx.getKeyInfoRepo().save(context, pair, keyInfo);
             }
         } catch (Exception e) {
@@ -198,9 +205,20 @@ public class ExpireOps {
         }
         String type = hashKey.substring(0, index);
         String key = hashKey.substring(index+1);
+        String traceId = parts[parts.length-1];
 
-        String traceId = parts[2];
-
+        String expireString = null;
+        String queryKey = null;
+        boolean noOps = false;
+        if (parts.length > 3) {
+            String ops = parts[2];
+            if (ops.startsWith("NOOPS")) { // NOOPS/300 or NOOPS=beanName/300
+                noOps = true;
+                String[] subParts = ops.split("/");
+                queryKey = subParts[0];
+                if (subParts.length > 1) expireString = subParts[1];
+            }
+        }
         Context context = new Context(traceId);
         KvPair pair = new KvPair(key, type);
 
@@ -225,16 +243,36 @@ public class ExpireOps {
 
             KvPairs pairs = new KvPairs(pair);
             AnyKey anyKey = new AnyKey();
+            KeyInfo keyInfo = null;
 
-            if (!AppCtx.getKeyInfoRepo().find(context, pairs, anyKey)) {
+            if (noOps) {
+
+                List<String> primaryIndexes = AppCtx.getDbaseOps().getPrimaryIndexes(context, type);
+                if (primaryIndexes == null || primaryIndexes.size() > 1) {
+                    String msg = "unsupported primary index cases for NOOPS for table: " + type;
+                    LOGGER.trace(msg);
+                    context.closeMonitor();
+                    return;
+                }
+                keyInfo = new KeyInfo(type, primaryIndexes.get(0), key);
+                if (expireString != null) {
+                    keyInfo.setExpire(expireString);
+                }
+                anyKey.add(keyInfo);
+
+            } else if (AppCtx.getKeyInfoRepo().find(context, pairs, anyKey)) {
+
+                keyInfo = anyKey.getKeyInfo();
+                queryKey = keyInfo.getQueryKey();
+
+            } else {
 
                 String msg = "keyInfo not found";
                 LOGGER.error(msg);
                 context.logTraceMessage(msg);
                 return;
-            }
 
-            KeyInfo keyInfo = anyKey.getKeyInfo();
+            }
 
             LOGGER.trace(keyInfo.toString());
 
@@ -243,11 +281,18 @@ public class ExpireOps {
             if (expire > 0) {
                 if (AppCtx.getRedisRepo().find(context, pairs, anyKey)) {
 
-                    String qkey = keyInfo.getQueryKey();
-                    if (qkey == null || !qkey.equals("NOOPS")) {
+                    String beanName = null;
+                    String[] ps = queryKey.split("=");
+                    if (ps.length > 1 && ps[1].length() > 0) {
+                        beanName = ps[1];
+                    }
+
+                    if (!noOps && beanName == null) {
+
                         AppCtx.getDbaseRepo().save(context, pairs, anyKey);
-                    } else if (qkey != null && qkey.startsWith("ExpireDbOps::")) {
-                        String beanName = qkey.substring(13);
+
+                    } else if (beanName != null) {
+
                         ApplicationContext ctx = AppCtx.getApplicationContext();
                         if (ctx != null) {
                             try {
@@ -261,11 +306,13 @@ public class ExpireOps {
                                 e.printStackTrace();
                             }
                         }
-                    } else {
-                        LOGGER.trace("queryKey = " + keyInfo.getQueryKey());
                     }
+
                     AppCtx.getRedisRepo().delete(context, pairs, anyKey);
-                    AppCtx.getKeyInfoRepo().delete(context, pairs);
+
+                    if (!noOps) {
+                        AppCtx.getKeyInfoRepo().delete(context, pairs);
+                    }
 
                 } else {
                     String msg = "failed to find key from redis for " + key;
